requirements.txt
numpy==1.23.5
neat-python==0.92
graphviz==0.20.1
scikit-learn==1.3.2
stockfish==3.28.0